Move monumentsCardList component to monument
